+++ src/main/java/com/configcat/ConfigCatClient.java
@@ -4,6 +4,9 @@ import com.configcat.evaluation.*;
 import com.configcat.log.ConfigCatLogMessages;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
+import com.configcat.override.OverrideBehaviour;
+import com.configcat.override.OverrideDataSource;
+import com.configcat.override.OverrideDataSourceBuilder;
 import com.google.gson.JsonElement;
 import okhttp3.OkHttpClient;
 import org.slf4j.LoggerFactory;
