+++ src/test/java/com/configcat/LocalTests.java
@@ -1,5 +1,7 @@
 package com.configcat;
 
+import com.configcat.override.OverrideBehaviour;
+import com.configcat.override.OverrideDataSourceBuilder;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import org.junit.jupiter.api.Test;
